+++ src/components/radio-button/radio-button-test.stories.tsx
@@ -4,18 +4,15 @@ import { StoryFn } from "@storybook/react";
 import { RadioButtonGroup, RadioButton } from ".";
 import { RadioButtonGroupProps } from "./radio-button-group/radio-button-group.component";
 import { RadioButtonProps } from "./radio-button.component";
-import CarbonProvider from "../carbon-provider";
 
 export default {
   title: "Radio Button/Test",
   includeStories: [
-    "WithLabelHelp",
+    "Required",
     "WithValidationsOnButtons",
     "WithValidationsOnRadioGroup",
     "WithTooltipPosition",
     "WithTooltipPositionOnRadioGroup",
-    "WithNewValidation",
-    "WithNewValidationGroup",
   ],
   parameters: {
     info: { disable: true },
@@ -23,52 +20,23 @@ export default {
       disableSnapshot: true,
     },
   },
-  argTypes: {
-    labelSpacing: {
-      options: [1, 2],
-      control: {
-        type: "select",
-      },
-    },
-    size: {
-      options: ["small", "large"],
-      control: {
-        type: "select",
-      },
-    },
-  },
 };
 
-export const WithLabelHelp: StoryFn<typeof RadioButton> = () => (
-  <RadioButtonGroup name="labelHelp" legend="Radio group legend">
-    <RadioButton
-      id="radio-1"
-      value="radio1"
-      label="Radio Option 1"
-      labelHelp="Radio 1"
-    />
-    <RadioButton
-      id="radio-2"
-      value="radio2"
-      label="Radio Option 2"
-      labelHelp="Radio 2"
-    />
-    <RadioButton
-      id="radio-3"
-      value="radio3"
-      label="Radio Option 3"
-      labelHelp="Radio 3"
-    />
+export const Required: StoryFn<typeof RadioButton> = () => (
+  <RadioButtonGroup name="required" legend="Radio group legend" required>
+    <RadioButton id="radio-1" value="radio1" label="Radio Option 1" />
+    <RadioButton id="radio-2" value="radio2" label="Radio Option 2" />
+    <RadioButton id="radio-3" value="radio3" label="Radio Option 3" />
   </RadioButtonGroup>
 );
 
-WithLabelHelp.storyName = "with labelHelp";
+Required.storyName = "required";
 
-export const WithValidationsOnButtons = ({ ...args }) => (
+export const WithValidationsOnButtons: StoryFn<typeof RadioButton> = () => (
   <RadioButtonGroup
     name="validations-on-buttons-group"
     onChange={() => console.log("change")}
-    {...args}
+    legend="Radio group legend"
   >
     <RadioButton
       id="validations-on-buttons-radio-1"
@@ -93,18 +61,13 @@ export const WithValidationsOnButtons = ({ ...args }) => (
 );
 
 WithValidationsOnButtons.storyName = "with validations on RadioButton";
-WithValidationsOnButtons.args = {
-  legend: "Radio group legend",
-  legendInline: false,
-  required: false,
-  inline: false,
-};
 
-export const WithValidationsOnRadioGroup = ({ ...args }) => (
+export const WithValidationsOnRadioGroup: StoryFn<typeof RadioButton> = () => (
   <RadioButtonGroup
     name="validations-on-group"
     onChange={() => console.log("change")}
-    {...args}
+    legend="Radio group legend"
+    error="Error message"
   >
     <RadioButton
       id="validations-on-group-radio-1"
@@ -125,14 +88,6 @@ export const WithValidationsOnRadioGroup = ({ ...args }) => (
 );
 
 WithValidationsOnRadioGroup.storyName = "with validations on RadioGroup";
-WithValidationsOnRadioGroup.args = {
-  error: "Error message",
-  warning: "",
-  legend: "Radio group legend",
-  legendInline: false,
-  required: false,
-  inline: false,
-};
 
 export const WithTooltipPosition: StoryFn<typeof RadioButton> = () => (
   <RadioButtonGroup
@@ -183,52 +138,52 @@ export const WithTooltipPositionOnRadioGroup: StoryFn<
 WithTooltipPositionOnRadioGroup.storyName =
   "with tooltip position on RadioGroup";
 
-export const WithNewValidation = (props: Partial<RadioButtonProps>) => {
+const radioContainerWidth = 400;
+
+export const RadioButtonComponent = (props: Partial<RadioButtonProps>) => {
+  const [isChecked, setIsChecked] = React.useState(false);
   return (
-    <CarbonProvider validationRedesignOptIn>
+    <div
+      style={{
+        marginTop: "64px",
+        marginLeft: "64px",
+        width: radioContainerWidth,
+      }}
+    >
       <RadioButton
         id="radio-1"
         value="radio1"
         label="Radiobutton 1"
+        checked={isChecked}
+        onChange={(e) => setIsChecked(e.target.checked)}
         {...props}
       />
-    </CarbonProvider>
+    </div>
   );
 };
 
-WithNewValidation.args = {
-  error: "Error message",
-  warning: "",
-  fieldHelp: "",
-  labelHelp: "",
-  required: false,
-  checked: false,
-  labelSpacing: 1,
-};
-
-export const WithNewValidationGroup = ({
+export const RadioButtonGroupComponent = ({
+  children,
   ...props
 }: Partial<RadioButtonGroupProps>) => {
   return (
-    <CarbonProvider validationRedesignOptIn>
+    <div
+      style={{
+        marginTop: "64px",
+        marginLeft: "64px",
+      }}
+    >
       <RadioButtonGroup
-        name="radio-button-group"
+        name="radiobuttongroup"
         legend="Radio group legend"
         {...props}
       >
         <RadioButton id="radio-1" value="radio1" label="Yes" />
         <RadioButton id="radio-2" value="radio2" label="No" />
         <RadioButton id="radio-3" value="radio3" label="Maybe" />
-      </RadioButtonGroup>
-    </CarbonProvider>
-  );
-};
 
-WithNewValidationGroup.args = {
-  error: "Error message",
-  warning: "",
-  legendHelp: "Legend help text",
-  legendInline: false,
-  required: true,
-  inline: false,
+        {children}
+      </RadioButtonGroup>
+    </div>
+  );
 };